% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_utiles.R
\name{distance}
\alias{distance}
\title{Distance formula}
\usage{
distance(votant, candidats)
}
\arguments{
\item{votant}{array}

\item{candidats}{array}
}
\value{
distance
}
\description{
Distance formula
}
