% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_utiles.R
\name{distance_to_pref}
\alias{distance_to_pref}
\title{Distance formula}
\usage{
distance_to_pref(distance_matrix)
}
\arguments{
\item{distance_matrix}{distance_matrix}
}
\value{
mat_inverse
}
\description{
Distance formula
}
