% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_utiles.R
\name{DistToScores}
\alias{DistToScores}
\title{Distance to score}
\usage{
DistToScores(dist, dim = 2, method = "linear", lambda = 5)
}
\arguments{
\item{dist}{int}

\item{dim}{dimension int}

\item{method}{method string}

\item{lambda}{lambdad int}
}
\value{
score
}
\description{
Distance to score
}
