% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_beta}
\alias{generate_beta}
\title{Generates a simulation of voting according to a beta law, returns voters preferences}
\usage{
generate_beta(
  n_voters,
  n_candidats,
  beta_a = 0.5,
  beta_b = 0.5,
  lambda = 0,
  min = 0,
  max = 1
)
}
\arguments{
\item{n_voters}{integer}

\item{n_candidats}{integer}

\item{beta_a}{double}

\item{beta_b}{double}

\item{lambda}{double}

\item{min}{int}

\item{max}{int}
}
\value{
scores
}
\description{
Generates a simulation of voting according to a beta law, returns voters preferences
}
