% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_binomial}
\alias{generate_binomial}
\title{Generate binomial scores}
\usage{
generate_binomial(n_voters, n_candidates, min = 0, max = 10, mean = 5)
}
\arguments{
\item{n_voters}{integer, the number of voters to generate scores for.}

\item{n_candidates}{integer,  The number of candidates to generate scores for.}

\item{min}{The minimum value of the  distribution, by default 0}

\item{max}{The maximum value of the  distribution, by default 10}

\item{mean}{The mean value of the  distribution, by default 5}
}
\value{
A matrix of scores with 'n_candidates' rows and 'n_voters' columns.
}
\description{
This function generates  discrete scores following a binomial distribution on a given scale
}
\examples{
voting_situation <- generate_binomial(n_voters=10, n_candidates=3, min=0, max=7, mean=5)
}
