% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_one_beta_two_unif_candidate}
\alias{generate_one_beta_two_unif_candidate}
\title{generate one beta and two unif candidate, returns voters preferences}
\usage{
generate_one_beta_two_unif_candidate(
  n_voters,
  n_candidats,
  beta_a = 0.5,
  beta_b = 0.5,
  lambda = 0,
  min = 0,
  max = 1
)
}
\arguments{
\item{n_voters}{integer}

\item{n_candidats}{integer}

\item{beta_a}{double}

\item{beta_b}{double}

\item{lambda}{double}

\item{min}{integer}

\item{max}{integer}
}
\value{
scores
}
\description{
generate one beta and two unif candidate, returns voters preferences
}
