% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_unif_continu}
\alias{generate_unif_continu}
\title{Generates a simulation of voting according to a uniform law, returns voters preferences}
\usage{
generate_unif_continu(n_voters, n_candidats, min = 0, max = 1)
}
\arguments{
\item{n_voters}{integer}

\item{n_candidats}{integer}

\item{min}{integer}

\item{max}{integer}
}
\value{
scores
}
\description{
Generates a simulation of voting according to a uniform law, returns voters preferences
}
