% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_unif_continuous}
\alias{generate_unif_continuous}
\title{Generates a simulation of voting according to a uniform law, returns voters preferences}
\usage{
generate_unif_continuous(n_voters, n_candidates, min = 0, max = 1)
}
\arguments{
\item{n_voters}{integer, represents the number of voters in the election}

\item{n_candidates}{integer, represents the number of candidates in the election}

\item{min}{int, the minimum value of the range of possible scores (by default 0)}

\item{max}{int, the maximum value of the range of possible scores (by default 1)}
}
\value{
scores
}
\description{
Generates a simulation of voting according to a uniform law, returns voters preferences
}
\examples{
voting_situation<- generate_unif_continuous(n_voters=10, n_candidates=3, min=0, max=10)
}
