% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_unif_disc}
\alias{generate_unif_disc}
\title{Generate uniform discrete scores}
\usage{
generate_unif_disc(n_voters, n_candidates, min = 0, max = 10)
}
\arguments{
\item{n_voters}{integer, the number of voters to generate scores for.}

\item{n_candidates}{integer,  The number of candidates to generate scores for.}

\item{min}{The minimum value of the  distribution, by default 0}

\item{max}{The maximum value of the  distribution, by default 10}
}
\value{
A matrix of scores with 'n_candidates' rows and 'n_voters' columns.
}
\description{
This function generates uniform discrete scores on a given scale
}
\examples{
voting_situation <- generate_unif_disc(n_voters=10, n_candidates=3, min=0, max=5)
}
