% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_utiles.R
\name{icdf}
\alias{icdf}
\title{Generalized inverse of the empirical cumulative function.}
\usage{
icdf(u, x, n)
}
\arguments{
\item{u}{a numeric vector of quantiles to be transformed.}

\item{x}{a numeric vector of data values.}

\item{n}{a positive integer specifying the length of the output vector.}
}
\value{
a numeric vector of transformed quantiles.
}
\description{
Generalized inverse of the empirical cumulative function.
}
\details{
Computes the generalized inverse of the empirical cumulative function,
which transforms quantiles \code{u} to the corresponding values of \code{x} based on
the frequency distribution of \code{x}.
}
