% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_utiles.R
\name{preferences_to_ranks}
\alias{preferences_to_ranks}
\title{Preferences_to_ranks}
\usage{
preferences_to_ranks(preferences)
}
\arguments{
\item{preferences}{voters preferences}
}
\value{
ranks
}
\description{
Preferences_to_ranks
}
