% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_utiles.R
\name{ScoresToDist}
\alias{ScoresToDist}
\title{Score to distance}
\usage{
ScoresToDist(x, dim = 2, method = "linear")
}
\arguments{
\item{x}{score}

\item{dim}{dimension int}

\item{method}{method string}
}
\value{
distance
}
\description{
Score to distance
}
