% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteSim-package.R
\docType{package}
\name{voteSim-package}
\alias{voteSim}
\alias{voteSim-package}
\title{voteSim: Generate Simulated Data for Voting Rules using Evaluations}
\description{
Provide functions to generate random simulated evaluations on candidates by voters for evaluation-based elections. Functions are based on several models for continuous or discrete evaluations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://eric.univ-lyon2.fr/arolland/}
}

}
\author{
\strong{Maintainer}: Antoine Rolland \email{antoine.rolland@univ-lyon2.fr}

Authors:
\itemize{
  \item Nathan Grimault
}

}
\keyword{internal}
